Timeout requests to local session after 100ms.
By creating the request to the local session on the heap,
rather than on the stack we can time it out without the risk of the
background thread accessing a request object that has been
deallocated.

To ensure correct ownership a new data structure is introduced -
a LeasedPtr.

Local requests cannot be issued concurrently.